+++ trunk/MAC/APL/PAC/CAL_Protocol/include/APL/CAL_Protocol/SubArray.h
@@ -70,7 +70,7 @@ namespace LOFAR {
 	       double                         sampling_frequency,
 	       int                            nyquist_zone,
 	       int                            nsubbands,
-	       uint8                          rcucontrol);
+	       uint32                         rcucontrol);
       virtual ~SubArray();
 
       /**
